// src/common/constants/deductions.enum.ts
export enum EDeductionns {
  Ciclica = "Ciclica",
  Eventuales = "Eventual",
}

export enum ERentDeductions {
  D01 = "Intereses crédito hipotecario",
  D02 = "Medicina prepagada",
  D03 = "Aportes a salud - contratista",
  D04 = "Aportes a pensión - contratista",
  D05 = "Arl - contratista",
}

export enum ERentDeductionsId {
  D01 = "D01",
  D02 = "D02",
  D03 = "D03",
  D04 = "D04",
  D05 = "D05",
}